Update mongoid dependency to version 8.1.0
This commit updates the mongoid dependency in Gemfile and mongoid-versioning.gemspec to use version 8.1.0. It also adds the new mongoid version in README.md and updates the test matrix in the GitHub Actions configuration to run the tests with the updated mongoid version. In addition, the version of the mongoid-versioning gem is updated to 5.0.0.